Changed back to float64

+++ baler/modules/models.py
@@ -10,15 +10,15 @@ class george_SAE(nn.Module):
         self.device = device
 
         # encoder
-        self.en1 = nn.Linear(n_features, 200, dtype=torch.float, device=device)
-        self.en2 = nn.Linear(200, 100, dtype=torch.float, device=device)
-        self.en3 = nn.Linear(100, 50, dtype=torch.float, device=device)
-        self.en4 = nn.Linear(50, z_dim, dtype=torch.float, device=device)
+        self.en1 = nn.Linear(n_features, 200, dtype=torch.float64, device=device)
+        self.en2 = nn.Linear(200, 100, dtype=torch.float64, device=device)
+        self.en3 = nn.Linear(100, 50, dtype=torch.float64, device=device)
+        self.en4 = nn.Linear(50, z_dim, dtype=torch.float64, device=device)
         # decoder
-        self.de1 = nn.Linear(z_dim, 50, dtype=torch.float, device=device)
-        self.de2 = nn.Linear(50, 100, dtype=torch.float, device=device)
-        self.de3 = nn.Linear(100, 200, dtype=torch.float, device=device)
-        self.de4 = nn.Linear(200, n_features, dtype=torch.float, device=device)
+        self.de1 = nn.Linear(z_dim, 50, dtype=torch.float64, device=device)
+        self.de2 = nn.Linear(50, 100, dtype=torch.float64, device=device)
+        self.de3 = nn.Linear(100, 200, dtype=torch.float64, device=device)
+        self.de4 = nn.Linear(200, n_features, dtype=torch.float64, device=device)
 
         self.n_features = n_features
         self.z_dim = z_dim
